requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.60.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.75.0
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.9.4
ml_dtypes==0.5.3
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.13.1
six==1.17.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
termcolor==3.1.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
zipp==3.23.0
